+++ apps/3d_rec_framework/src/pipeline/global_nn_recognizer_cvfh.cpp
@@ -9,27 +9,27 @@
 #include "pcl/apps/3d_rec_framework/utils/metrics.h"
 
 POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Histogram<367>,
-    (float[367], histogram, histogram367)
+    (double[367], histogram, histogram367)
 )
 
 POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Histogram<431>,
-    (float[431], histogram, histogram431)
+    (double[431], histogram, histogram431)
 )
 
 POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Histogram<559>,
-    (float[559], histogram, histogram559)
+    (double[559], histogram, histogram559)
 )
 
 POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Histogram<815>,
-    (float[815], histogram, histogram815)
+    (double[815], histogram, histogram815)
 )
 
 POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Histogram<879>,
-    (float[879], histogram, histogram879)
+    (double[879], histogram, histogram879)
 )
 
 POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Histogram<1327>,
-    (float[1327], histogram, histogram1327)
+    (double[1327], histogram, histogram1327)
 )
